pyproject.toml
[project]
name = "llama-agent"
version = "0.1.0"
description = "Llama Agent with FastAPI backend and NextJS frontend"
authors = [
    {name = "nerifuture", email = "notintricate@users.noreply.github.com"},
]
dependencies = [
    "fastapi>=0.109.1",
    "uvicorn[standard]>=0.23.2",
    "python-dotenv>=1.0.0",
    "aiostream>=0.6.4",
    "colorama>=0.4.6",
    "qdrant-client>=1.12.1",
    "llama-index-core>=0.12.3",
    "llama-index>=0.12.3",
    "llama-index-llms-perplexity>=0.3.1",
    "llama-index-vector-stores-qdrant>=0.4.0",
 
 

]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "python run.py dev"
prod = "python run.py prod"

[tool.pdm]
distribution = false
